+++ packages/core/dialect/mysql/default.ts
@@ -15,6 +15,9 @@ export default class MySQLDefault extends GenericDialect<MySQLLib.Pool> implemen
     }
 
     const { ssl } = this.credentials.mysqlOptions || <ConnectionInterface['mysqlOptions']>{};
+    if (typeof ssl === 'boolean') {
+      throw new Error('SSL as boolean only supported for xprotocol. See: https://mtxr.gitbook.io/vscode-sqltools/connections/mysql#2-ssl');
+    }
 
     const pool = MySQLLib.createPool({
       connectTimeout: this.credentials.connectionTimeout * 1000,